separate Canvas module
