+++ src/Canvas/Drawable/DrawableCircle.test.js
