+++ src/Canvas/Drawable/DrawableRectangle.js
