+++ src/Canvas/Wall.js
